requirements.txt
beautifulsoup4==4.12.3
discord.py==2.3.2
playwright==1.41.0
python-dotenv==1.0.1
